requirements.txt
pydantic
fastapi
uvicorn
requests
asyncio
python-dotenv
boto3
pandas
numpy
docling
mistralai
webdriver-manager
selenium
pinecone
chromadb
scikit-learn